+++ src/main/java/io/junrock/GAZA/global/type/MessageType.java
@@ -1,11 +1,11 @@
-package io.junrock.GAZA.domain.message.dto;
+package io.junrock.GAZA.global.type;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public enum TypeMessage {
+public enum MessageType {
     MESSAGE("message"),
     DONATE("donate");
     private String donateType;
